api/requirements.txt
bcrypt==3.2.0
blinker==1.4
boto3==1.20.47
botocore==1.23.47
Bottleneck==1.3.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
lightgbm==3.3.2
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.27.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
numexpr==2.8.1
numpy==1.21.5
packaging==21.3
pandas==1.3.4
pip==21.2.4
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.4
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
s3transfer==0.5.1
scikit-learn==1.0.2
scipy==1.7.3
setuptools==58.0.4
six==1.16.0
SQLAlchemy==1.4.31
threadpoolctl==3.1.0
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
wheel==0.37.1
wincertstore==0.2
zipp==3.7.0
